# api/swagger.yaml
basePath: /api
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        description: Error Message
        type: string
      status:
        description: Statustext of Errorcode
        type: string
    type: object
  api.FileStashUrl:
    properties:
      url:
        type: string
    type: object
  api.InfluxdbConfiguration:
    properties:
      batch_size:
        type: integer
      database_name:
        type: string
      host:
        type: string
      max_retries:
        type: integer
      max_retry_time:
        type: string
      meta_as_tags:
        type: string
      organization:
        type: string
      password:
        type: string
      port:
        type: integer
      retry_exponential_base:
        type: integer
      retry_interval:
        type: string
      ssl_enabled:
        type: boolean
      type:
        type: string
      user:
        type: string
    type: object
  api.LogicalVolume:
    properties:
      lv_attr:
        type: string
      lv_name:
        type: string
      lv_size:
        type: string
      machine_id:
        type: string
      vg_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@clustercockpit.org
    name: ClusterCockpit Project
    url: https://github.com/Deepbinder-main
  description: API for batch job control.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: ClusterCockpit REST API
  version: 1.0.0
paths:
  /file_stash_url:
    delete:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Deletes the File Stash URL
      tags:
      - FileStash
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved File Stash URL
          schema:
            $ref: '#/definitions/api.FileStashUrl'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Retrieves the File Stash URL
      tags:
      - FileStash
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File Stash URL
        in: formData
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Created/Updated File Stash URL
          schema:
            $ref: '#/definitions/api.FileStashUrl'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Creates or updates File Stash URL
      tags:
      - FileStash
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: File Stash URL
        in: formData
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated File Stash URL
          schema:
            $ref: '#/definitions/api.FileStashUrl'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Updates the File Stash URL
      tags:
      - FileStash
  /influxdb_config:
    delete:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Deletes the InfluxDB configuration
      tags:
      - InfluxDB
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved InfluxDB configuration
          schema:
            $ref: '#/definitions/api.InfluxdbConfiguration'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Retrieves the InfluxDB configuration
      tags:
      - InfluxDB
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Type
        in: formData
        name: type
        required: true
        type: string
      - description: Database Name
        in: formData
        name: database_name
        required: true
        type: string
      - description: Host
        in: formData
        name: host
        required: true
        type: string
      - description: Port
        in: formData
        name: port
        required: true
        type: integer
      - description: User
        in: formData
        name: user
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Organization
        in: formData
        name: organization
        required: true
        type: string
      - description: SSL Enabled
        in: formData
        name: ssl_enabled
        required: true
        type: boolean
      - description: Batch Size
        in: formData
        name: batch_size
        required: true
        type: integer
      - description: Retry Interval
        in: formData
        name: retry_interval
        required: true
        type: string
      - description: Retry Exponential Base
        in: formData
        name: retry_exponential_base
        required: true
        type: integer
      - description: Max Retries
        in: formData
        name: max_retries
        required: true
        type: integer
      - description: Max Retry Time
        in: formData
        name: max_retry_time
        required: true
        type: string
      - description: Meta as Tags
        in: formData
        name: meta_as_tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Created/Updated InfluxDB configuration
          schema:
            $ref: '#/definitions/api.InfluxdbConfiguration'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Creates or updates InfluxDB configuration
      tags:
      - InfluxDB
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Type
        in: formData
        name: type
        required: true
        type: string
      - description: Database Name
        in: formData
        name: database_name
        required: true
        type: string
      - description: Host
        in: formData
        name: host
        required: true
        type: string
      - description: Port
        in: formData
        name: port
        required: true
        type: integer
      - description: User
        in: formData
        name: user
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Organization
        in: formData
        name: organization
        required: true
        type: string
      - description: SSL Enabled
        in: formData
        name: ssl_enabled
        required: true
        type: boolean
      - description: Batch Size
        in: formData
        name: batch_size
        required: true
        type: integer
      - description: Retry Interval
        in: formData
        name: retry_interval
        required: true
        type: string
      - description: Retry Exponential Base
        in: formData
        name: retry_exponential_base
        required: true
        type: integer
      - description: Max Retries
        in: formData
        name: max_retries
        required: true
        type: integer
      - description: Max Retry Time
        in: formData
        name: max_retry_time
        required: true
        type: string
      - description: Meta as Tags
        in: formData
        name: meta_as_tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated InfluxDB configuration
          schema:
            $ref: '#/definitions/api.InfluxdbConfiguration'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Updates the InfluxDB configuration
      tags:
      - InfluxDB
  /logical_volume:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Machine ID
        in: formData
        name: machine_id
        required: true
        type: string
      - description: Logical Volume Name
        in: formData
        name: lv_name
        required: true
        type: string
      - description: Volume Group Name
        in: formData
        name: vg_name
        required: true
        type: string
      - description: Logical Volume Attributes
        in: formData
        name: lv_attr
        required: true
        type: string
      - description: Logical Volume Size
        in: formData
        name: lv_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created logical volume
          schema:
            $ref: '#/definitions/api.LogicalVolume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Creates a new logical volume record
      tags:
      - LogicalVolume
  /logical_volume/{lv_id}:
    delete:
      parameters:
      - description: Logical Volume ID
        in: path
        name: lv_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Deletes a logical volume record
      tags:
      - LogicalVolume
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Logical Volume ID
        in: path
        name: lv_id
        required: true
        type: integer
      - description: Logical Volume Name
        in: formData
        name: lv_name
        required: true
        type: string
      - description: Volume Group Name
        in: formData
        name: vg_name
        required: true
        type: string
      - description: Logical Volume Attributes
        in: formData
        name: lv_attr
        required: true
        type: string
      - description: Logical Volume Size
        in: formData
        name: lv_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated logical volume
          schema:
            $ref: '#/definitions/api.LogicalVolume'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Updates a logical volume record
      tags:
      - LogicalVolume
  /logical_volumes/{machine_id}:
    get:
      parameters:
      - description: Machine ID
        in: path
        name: machine_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved logical volumes
          schema:
            items:
              $ref: '#/definitions/api.LogicalVolume'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Retrieves logical volume records for a machine
      tags:
      - LogicalVolume
...
  /machines:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of machines
          schema:
            items:
              $ref: '#/definitions/api.Machine'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Lists all machines
      tags:
      - Machine
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Auth-Token
    type: apiKey
swagger: "2.0"